refactor: replace inline NodeClient modules with cardano-node-clients library
Delete 6 inline modules (NodeClient/*, Core/Balance) totalling ~780 lines
and replace with imports from the cardano-node-clients library. Provider
and Submitter adapters delegate to the library, keeping MPFS interface
types unchanged.
