+++ cabal.project
@@ -54,5 +54,10 @@ source-repository-package
   tag: 9c2555a5e6d72fdd55eb1e13cff812d72ea41913
   --sha256: 03s9q9arz8km5hc08zsv7zxjzigy34rqf521p94xm2n4vmly2i7f
 
+source-repository-package
+  type: git
+  location: https://github.com/paolino/cardano-node-clients
+  tag: 7677065cbcadaa784e60e6974b552a9cd7ef0dc3
+
 -- Enable iohk flag for contra-tracer-contrib to use CHaP's simple API
 constraints: contra-tracer-contrib +iohk
