+++ cardano-mpfs-offchain/lib/Cardano/MPFS/Provider/NodeClient.hs
@@ -1,44 +1,21 @@
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE PatternSynonyms #-}
-
 -- |
 -- Module      : Cardano.MPFS.Provider.NodeClient
 -- Description : N2C-backed Provider via LocalStateQuery
 -- License     : Apache-2.0
 --
 -- Production implementation of the 'Provider'
--- interface. Queries a local Cardano node via the
--- N2C LocalStateQuery mini-protocol using an
--- 'LSQChannel' (see "Cardano.MPFS.NodeClient.Connection").
---
--- Protocol parameters and UTxOs are retrieved with
--- Conway-era queries. An era mismatch (node not yet
--- in Conway) results in a runtime error.
+-- interface. Delegates to the @cardano-node-clients@
+-- library for N2C LocalStateQuery queries, adding
+-- the MPFS-specific 'evaluateTx' stub.
 module Cardano.MPFS.Provider.NodeClient
     ( -- * Construction
       mkNodeClientProvider
     ) where
 
-import Data.Map.Strict qualified as Map
-import Data.Set qualified as Set
+import Cardano.Node.Client.N2C.Provider qualified as Lib
+import Cardano.Node.Client.N2C.Types (LSQChannel)
+import Cardano.Node.Client.Provider qualified as Lib
 
-import Cardano.Ledger.State (UTxO (..))
-
-import Ouroboros.Consensus.Cardano.Block
-    ( pattern QueryIfCurrentConway
-    )
-import Ouroboros.Consensus.Ledger.Query
-    ( Query (BlockQuery)
-    )
-import Ouroboros.Consensus.Shelley.Ledger.Query
-    ( pattern GetCurrentPParams
-    , pattern GetUTxOByAddress
-    )
-
-import Cardano.MPFS.NodeClient.LocalStateQuery
-    ( queryLSQ
-    )
-import Cardano.MPFS.NodeClient.Types (LSQChannel)
 import Cardano.MPFS.Provider (Provider (..))
 
 -- | Create a 'Provider IO' backed by the N2C
@@ -48,37 +25,14 @@ mkNodeClientProvider
     -- ^ LocalStateQuery channel to the Cardano node
     -> Provider IO
 mkNodeClientProvider ch =
-    Provider
-        { queryProtocolParams = do
-            result <-
-                queryLSQ ch
-                    $ BlockQuery
-                    $ QueryIfCurrentConway
-                        GetCurrentPParams
-            case result of
-                Right pp -> pure pp
-                Left _mismatch ->
-                    error
-                        "queryProtocolParams: era \
-                        \mismatch — node not in Conway"
-        , queryUTxOs = \addr -> do
-            result <-
-                queryLSQ ch
-                    $ BlockQuery
-                    $ QueryIfCurrentConway
-                    $ GetUTxOByAddress
-                        (Set.singleton addr)
-            case result of
-                Right utxo ->
-                    pure
-                        $ Map.toList
-                        $ unUTxO utxo
-                Left _mismatch ->
-                    error
-                        "queryUTxOs: era mismatch \
-                        \— node not in Conway"
-        , evaluateTx = \_ ->
-            error
-                "evaluateTx: requires local Plutus \
-                \evaluator"
-        }
+    let libProv = Lib.mkN2CProvider ch
+    in  Provider
+            { queryProtocolParams =
+                Lib.queryProtocolParams libProv
+            , queryUTxOs =
+                Lib.queryUTxOs libProv
+            , evaluateTx = \_ ->
+                error
+                    "evaluateTx: requires local \
+                    \Plutus evaluator"
+            }
