+++ cardano-mpfs-offchain/lib/Cardano/MPFS/TxBuilder/Real/End.hs
@@ -48,7 +48,6 @@ import PlutusTx.Builtins.Internal
     ( BuiltinByteString (..)
     )
 
-import Cardano.MPFS.Core.Balance (balanceTx)
 import Cardano.MPFS.Core.OnChain
     ( CageDatum (..)
     , MintRedeemer (..)
@@ -64,6 +63,7 @@ import Cardano.MPFS.TxBuilder.Config
     ( CageConfig (..)
     )
 import Cardano.MPFS.TxBuilder.Real.Internal
+import Cardano.Node.Client.Balance (balanceTx)
 
 import Data.List (sortOn)
 import Data.Ord (Down (..))
