+++ cardano-mpfs-offchain/lib/Cardano/MPFS/TxBuilder/Real/Retract.hs
@@ -58,7 +58,6 @@ import Cardano.Ledger.Conway.Scripts
 import Cardano.Ledger.Core (hashScript)
 import Cardano.Ledger.TxIn (TxIn)
 
-import Cardano.MPFS.Core.Balance (balanceTx)
 import Cardano.MPFS.Core.OnChain
     ( CageDatum (..)
     , OnChainRequest (..)
@@ -78,6 +77,7 @@ import Cardano.MPFS.TxBuilder.Config
     ( CageConfig (..)
     )
 import Cardano.MPFS.TxBuilder.Real.Internal
+import Cardano.Node.Client.Balance (balanceTx)
 import PlutusTx.Builtins.Internal
     ( BuiltinByteString (..)
     )
